move content stream code to own file
